cloud_function/copy_tables_between_bq_dataset/requirements.txt
# Function dependencies, for example:
# package>=version

google-cloud-bigquery
